Rename hackrf-ui.[ch]
Every other source file uses underscores! :(
