+++ firmware/common/ui_portapack.h
@@ -22,7 +22,7 @@
 #ifndef __UI_PORTAPACK_H__
 #define __UI_PORTAPACK_H__
 
-#include "hackrf-ui.h"
+#include "hackrf_ui.h"
 
 const hackrf_ui_t* portapack_hackrf_ui_init() __attribute__((weak));
