+++ firmware/hackrf-common.cmake
@@ -159,7 +159,7 @@ macro(DeclareTargets)
 		${PATH_HACKRF_FIRMWARE_COMMON}/spi_bus.c
 		${PATH_HACKRF_FIRMWARE_COMMON}/spi_ssp.c
 		${PATH_HACKRF_FIRMWARE_COMMON}/gpio_lpc.c
-		${PATH_HACKRF_FIRMWARE_COMMON}/hackrf-ui.c
+		${PATH_HACKRF_FIRMWARE_COMMON}/hackrf_ui.c
 	)
 
 	if(BOARD STREQUAL "RAD1O")
